Make mapper package private
